refactor: Abstract search interface
